ACTIN-1570: Update SERVE SQL for v7 datamodel (WIP)
